Improved the descriptions and comments for the Object Detection example.
