+++ xt_training/od_lib/__init__.py
@@ -4,3 +4,5 @@
 #        COPPE - Universidade Federal do Rio de Janeiro                                   #
 #        Last modification: May 24th 2018                                                 #
 ###########################################################################################
+
+__all__ = ["BoundingBox", "Evaluator", "utils"]
